Hide internalBatchRequest
